requirements.txt
python-dotenv~=1.0.0
pymongo~=4.4.1
pydantic~=1.10.12
fastapi~=0.101.1
uvicorn~=0.23.2
werkzeug
starlette
PyJWT
python-jose[cryptography]
passlib[bcrypt]~=1.7.4
python-multipart
py3-validate-email
cryptography
requests~=2.31.0
jose~=1.0.0
requests
httpx
pandas
openpyxl
pdfkit